Update README.md
Overview
Here, we quantify the relationship between changes in the extinction risk of 4,426 terrestrial mammals over a 24-year period (1996-2020), the fragmentation of their suitable habitat, and the levels of human pressure within the associated habitat matrix. For each species, we quantify the degree of fragmentation as the average of the Euclidean distance of all patches of suitable habitat from the nearest patch edge, the degree of patch isolation as the average of the Euclidean distance of the area of ‘unsuitable’ habitat from the nearest edge, and matrix condition as the extent and change over time of high human pressure levels overlapping with the surrounding matrix. Spatial data representing the condition of the matrix were obtained from the recently updated human footprint maps, which provide a single metric that combines data on human activities, all of which are driving the current biodiversity crisis. We define a human footprint threshold of ≥ 3 out of 50 to represent the extent of human-modified habitat within the matrix. Following previous studies42,43,54, we classify species into two groups of extinction risk, low-risk transitions and high-risk transitions (Fig. 1), based on the first and last Red List category registered between 1996 and 2020. In combination with other predictors of extinction risk (see Table 1 for a description), we quantify the relative importance of habitat fragmentation and the condition of the matrix for determining extinction risk transitions in terrestrial mammals.